requirements.txt
openai
python-dotenv
shapely